Added support for left analog stick control (e.g. Oculus Arcade)
By default the PinSim joystick maps to the Xbox 360 D-pad. To map it to
the left analog stick instead, hold "Back" and presss "Up" on the
joystick. To revert to D-pad, hold "Back" and press "Down" on the
joystick. Alternatively, booting while holding the right flipper will
enable this mode (in addition to disabling the accelerometer like
before).